+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # koda-times change Logs
 
 
+# 1.1.0
+
+*   Add arithmatic operation for Period, Instace, LocalDateTime, LocalDate, LocalTime
+*   Remove DurationBuilder class
+* Upgrade kotlin library to 1.1 
+
 # 1.0.0
 
 ### First release
 
--. Supply joda-time DateTime extension functions
+*   Supply joda-time DateTime extension functions